Refactor: api코드 분리
