+++ app/(afterLogin)/estimate/@estimateTrainer/components/MyEstimateU.tsx
@@ -7,6 +7,8 @@ import { sendRequest } from "@/app/api/rootApi";
 import { useUserLocation } from "@/store/useUserLocation";
 import { useEffect } from "react";
 import { EstimateItem } from "../../types";
+import { getImmeEstimateU } from "../../api/getImmeEstimateU";
+import { getMyEstimateT } from "../../api/getMyEstimateT";
 
 export interface MyEstimatePropsT {
   data: {
@@ -17,29 +19,18 @@ export interface MyEstimatePropsT {
     ptProgram: { id: number; ptProgramName: string };
   };
 }
-const immeEstimateU = async () => {
-  return await sendRequest(
-    `trainer/estimates?courseId=0&size=5&ptAddress=서울특별시`,
-    "get",
-  );
-};
-const fetchMyEstimateT = (id: number | undefined | null) => async () => {
-  if (id) {
-    return await sendRequest(`trainer/estimates/${id}`);
-  }
-};
 
 export default function MyEstimateU() {
   const { location } = useUserLocation();
 
   const userDataT = useQuery<EstimateItem>({
     queryKey: ["immeEstimateU"],
-    queryFn: immeEstimateU,
+    queryFn: getImmeEstimateU,
     enabled: !!location,
   });
   const { data, isLoading, refetch } = useQuery<MyEstimatePropsT>({
     queryKey: ["myEstimateT"],
-    queryFn: fetchMyEstimateT(userDataT.data?.myEstimateId),
+    queryFn: getMyEstimateT(userDataT.data?.myEstimateId),
     enabled: !!userDataT.data?.myEstimateId,
   });
