+++ app/(afterLogin)/estimate/@estimateUser/components/MyEstimateT.tsx
@@ -3,7 +3,7 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import BlankEstimate from "./BlankEstimate";
 import { getEstimates } from "../../types";
-import { sendRequest } from "@/app/api/rootApi";
+import { getMyEstimateU } from "../../api/getMtEstimateU";
 import MyEstimateCard from "./MyEstimateCard";
 
 export interface MyEstimatePropsU {
@@ -14,11 +14,6 @@ export interface MyEstimatePropsU {
     ptAddress: string;
   };
 }
-const fetchMyEstimateU = (id: number | undefined | null) => async () => {
-  if (id) {
-    return await sendRequest(`user/estimates/${id}`);
-  }
-};
 
 export default function MyEstimateT() {
   const queryClient = useQueryClient();
@@ -27,7 +22,7 @@ export default function MyEstimateT() {
 
   const { data, isLoading } = useQuery<MyEstimatePropsU>({
     queryKey: ["myEstimateU"],
-    queryFn: fetchMyEstimateU(trainerEstimates?.pages[0].myEstimateId),
+    queryFn: getMyEstimateU(trainerEstimates?.pages[0].myEstimateId),
     enabled: !!trainerEstimates?.pages[0].myEstimateId,
   });
