+++ app/(afterLogin)/estimate/api/getMyEstimateT.ts
@@ -1,0 +1,6 @@
+import { sendRequest } from "@/app/api/rootApi";
+export const getMyEstimateT = (id: number | undefined | null) => async () => {
+  if (id) {
+    return await sendRequest(`trainer/estimates/${id}`);
+  }
+};
