+++ app/(afterLogin)/estimate/api/getOneEstimateT.ts
@@ -1,0 +1,9 @@
+import { sendRequest } from "@/app/api/rootApi";
+export const getOneEstimateT = (dataId: number) => async () => {
+  let estimateT = null;
+
+  if (dataId) {
+    estimateT = await sendRequest(`trainer/estimates/${dataId}`);
+  }
+  return estimateT;
+};
