+++ app/(afterLogin)/estimatePaper/api/getOnlyTrainerPrograms.ts
@@ -1,0 +1,5 @@
+import { sendRequest } from "@/app/api/rootApi";
+
+export const getOnlyTrainerPrograms = async () => {
+  return await sendRequest("trainer/estimates/programs");
+};
